requirements.txt
# ============================================================================
# QUANTUM TRADER PRO - Requirements
# ============================================================================

# Core Trading
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
lightgbm>=4.0.0
optuna>=3.3.0  # Hyperparameter optimization

# Data & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async & Performance
aiohttp>=3.8.0
asyncio>=3.4.3
numba>=0.57.0  # JIT compilation

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Utilities
loguru>=0.7.0
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.5.0  # Beautiful terminal output

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Security
cryptography>=41.0.0
requests>=2.31.0  # For RSA testnet client

# Optional: GPU acceleration
#torch>=2.0.0  # Uncomment if you have CUDA
#torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
